add signals for profile and user connections
