+++ members/apps.py
@@ -4,3 +4,6 @@ from django.apps import AppConfig
 class MembersConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "members"
+
+    def ready(self):
+        import members.signals
