+++ members/models.py
@@ -1,17 +1,19 @@
 from django.db import models
 from django.contrib.auth.models import User
 import uuid
-# # Create your models here.
 
 from django.db.models.signals import post_save, post_delete
 from django.dispatch import receiver
+# # Create your models here.
+
+
 
 
 class Profile(models.Model):
     member = models.OneToOneField(User, on_delete=models.CASCADE, null=True, blank=True)
     name = models.CharField(max_length=200, blank=True, null=True)
     email = models.EmailField(max_length=500, blank=True, null=True)
-    membername = models.CharField(max_length=200, blank=True, null=True)
+    username = models.CharField(max_length=200, blank=True, null=True)
     location = models.CharField(max_length=200, blank=True, null=True)
     short_intro = models.CharField(max_length=200, blank=True, null=True)
     bio = models.TextField(blank=True, null=True)
@@ -27,7 +29,7 @@ class Profile(models.Model):
     id = models.UUIDField(default=uuid.uuid4, unique=True, primary_key=True, editable=False)
 
     def __str__(self):
-        return str(self.membername)
+        return str(self.username)
 
     class Meta:
         ordering = ["created"]
@@ -49,7 +51,7 @@ class Skill(models.Model):
     id = models.UUIDField(default=uuid.uuid4, unique=True, primary_key=True, editable=False)
 
     def __str__(self):
-        return str(self.membername)
+        return str(self.name)
 
 
 # class Message(models.Model):
@@ -71,20 +73,3 @@ class Skill(models.Model):
 
 #     class Meta:
 #         ordering = ['is_read', '-created']
-
-#@receiver(post_save, sender=Profile)
-def createProfile(sender, instance, created, **kwargs):
-    if created:
-        member = instance
-        profile = Profile.objects.create(
-            member=member,
-            membername=member.name,
-            email=member.email,
-            name=member.first_name,
-            )
-
-def deleteMember(sender, instance, **kwargs):
-    print('Deleting member...')
-
-post_save.connect(createProfile, sender=Profile)
-post_delete.connect(deleteMember, sender=Profile)
